Clean up StripeError.generate
