+++ lib/Error.js
@@ -12,6 +12,7 @@ class StripeError extends Error {
     this.raw = raw;
     this.rawType = raw.type;
     this.code = raw.code;
+    this.doc_url = raw.doc_url;
     this.param = raw.param;
     this.detail = raw.detail;
     this.headers = raw.headers;
@@ -38,6 +39,10 @@ class StripeError extends Error {
         return new StripeInvalidRequestError(rawStripeError);
       case 'api_error':
         return new StripeAPIError(rawStripeError);
+      case 'authentication_error':
+        return new StripeAuthenticationError(rawStripeError);
+      case 'rate_limit_error':
+        return new StripeRateLimitError(rawStripeError);
       case 'idempotency_error':
         return new StripeIdempotencyError(rawStripeError);
       case 'invalid_grant':
@@ -116,6 +121,7 @@ class StripeIdempotencyError extends StripeError {}
  */
 class StripeInvalidGrantError extends StripeError {}
 
+module.exports.generate = StripeError.generate;
 module.exports.StripeError = StripeError;
 module.exports.StripeCardError = StripeCardError;
 module.exports.StripeInvalidRequestError = StripeInvalidRequestError;
