+++ types/test/typescriptTest.ts
@@ -132,3 +132,9 @@ stripe.setHost('host', 'port', 'protocol');
     }
   }
 })();
+
+const stripeCardError: Stripe.StripeCardError = Stripe.errors.generate({
+  type: 'card_error',
+  code: 'card_declined',
+  charge: 'ch_123',
+});
